+++ .github/workflows/tests.yaml
@@ -11,7 +11,7 @@ jobs:
         uses: actions/checkout@v2
       - name: install poetry
         run:  curl -sSL https://raw.githubusercontent.com/python-poetry/poetry/master/install-poetry.py | python -
-      - name: install testix
+      - name: install test_first
         run:  poetry install
       - name: unit tests
         run:  poetry run rake units
@@ -22,7 +22,7 @@ jobs:
         uses: actions/checkout@v2
       - name: install poetry
         run:  curl -sSL https://raw.githubusercontent.com/python-poetry/poetry/master/install-poetry.py | python -
-      - name: install testix
+      - name: install test_first
         run:  poetry install
       - name: examples
         run:  poetry run rake examples
