+++ examples/tests/test_calculator.py
@@ -1,6 +1,6 @@
 import pytest
-from testix.frequentlyused import *
-from testix import patch_module
+from test_first.frequentlyused import *
+from test_first import patch_module
 from examples import calculator
 
 class Test_Calculator:
