+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
-name = "testix"
+name = "test_first"
 version = "8.0.1"
-description = "Mocking framework Python with *exact* Scenarios"
+description = "Test First Python Mocking Framework designed especially for TDD"
 authors = ["Yoav Kleinberger <haarcuba@gmail.com>"]
 license = "MIT"
 classifiers = [
