--- test/conftest.py
@@ -1,3 +1,0 @@
-from testix import failhooks
-
-failhooks.setMode( 'raise' )
