+++ test_first/fakefile.py
@@ -1,4 +1,4 @@
-from testix import fake
+from test_first import fake
 
 
 class FakeFile(fake.Fake):
