+++ test_first/frequentlyused.py
@@ -1,0 +1,6 @@
+from test_first.fake import Fake
+from test_first.fakeobject import FakeObject
+from test_first.argumentexpectations import *
+from test_first.scenario import *
+from test_first.hook import Hook
+from test_first.DSL import *
