--- testix.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-python testix.egg $@
