--- testix/frequentlyused.py
@@ -1,6 +1,0 @@
-from testix.fake import Fake
-from testix.fakeobject import FakeObject
-from testix.argumentexpectations import *
-from testix.scenario import *
-from testix.hook import Hook
-from testix.DSL import *
