+++ unit_tests/conftest.py
@@ -1,0 +1,3 @@
+from test_first import failhooks
+
+failhooks.setMode( 'raise' )
